feat(npd): expose CLI via NAPI runCli + ship npd bin shim
